Refactor package structure by moving bearertoken and httpfs to internal directory and adding new context and test files
